refactor: extract webhook delivery logic into separate package
